home: Add home page for non-logged in users
Previously, home page directly redirected to the login page. This commit adds a
home page that is visible to users even without being signed in. The previous
home page has now been moved to a dashboard page.

Closes #67
